strip out all non-json
