+++ python/cli/prelude_cli/views/detect.py
@@ -1,14 +1,10 @@
 import click
 
 from rich import print_json
-from rich.table import Table
-from rich.console import Console
-from collections import defaultdict
 from datetime import datetime, timedelta, time
 
+from prelude_sdk.models.codes import RunCode
 from prelude_cli.views.shared import handle_api_error
-from prelude_sdk.models.codes import RunCode, ExitCode
-from prelude_sdk.controllers.build_controller import BuildController
 from prelude_sdk.controllers.detect_controller import DetectController
 
 
@@ -27,7 +23,15 @@ def detect(ctx):
 def register_endpoint(controller, name, tags):
     """ Register a new endpoint """
     token = controller.register_endpoint(name=name, tags=tags)
-    click.secho(f'Your token: {token}', fg='green')
+    click.secho(token)
+
+
+@detect.command('tests')
+@click.pass_obj
+@handle_api_error
+def list_tests(controller):
+    """ List all security tests """
+    print_json(data=controller.list_tests())
 
 
 @detect.command('enable-test')
@@ -40,7 +44,7 @@ def register_endpoint(controller, name, tags):
 @click.pass_obj
 @handle_api_error
 def activate_test(controller, test, run_code, tags):
-    """ Add TEST to your queue """
+    """ Add test to your queue """
     controller.enable_test(ident=test, run_code=RunCode[run_code.upper()].value, tags=tags)
 
 
@@ -50,9 +54,8 @@ def activate_test(controller, test, run_code, tags):
 @click.pass_obj
 @handle_api_error
 def deactivate_test(controller, test):
-    """ Remove TEST from your queue """
+    """ Remove test from your queue """
     controller.disable_test(ident=test)
-    click.secho(f'Disabled {test}', fg='green')
 
 
 @detect.command('delete-endpoint')
@@ -63,7 +66,6 @@ def deactivate_test(controller, test):
 def delete_endpoint(controller, endpoint_id):
     """Delete a probe/endpoint"""
     controller.delete_endpoint(ident=endpoint_id)
-    click.secho(f'Deleted {endpoint_id}', fg='green')
 
 
 @detect.command('queue')
@@ -71,13 +73,7 @@ def delete_endpoint(controller, endpoint_id):
 @handle_api_error
 def queue(controller):
     """ List all tests in your active queue """
-    build = BuildController(account=controller.account)
-    tests = {row['id']: row['name'] for row in build.list_tests()}
-    active = controller.list_queue()
-    for q in active:
-        q['run_code'] = RunCode(q['run_code']).name
-        q['name'] = tests[q['test']]
-    print_json(data=active)
+    print_json(data=controller.list_queue())
 
 
 @detect.command('search')
@@ -105,11 +101,7 @@ def endpoints(controller):
 @handle_api_error
 def social_statistics(controller, test, days):
     """ Pull social statistics for a specific test """
-    stats = defaultdict(lambda: defaultdict(int))
-    for dos, values in controller.social_stats(ident=test, days=days).items():
-        for state, count in values.items():
-            stats[dos][ExitCode(int(state)).name] = count
-    print_json(data=stats)
+    print_json(data=controller.social_stats(ident=test, days=days).items())
 
 
 @detect.command('recommendations')
@@ -128,7 +120,6 @@ def recommendation(controller):
 def add_recommendation(controller, title, description):
     """ Create a new security recommendation """
     controller.create_recommendation(title=title, description=description)
-    click.secho('Successfully submitted recommendation', fg='green')
 
 
 @detect.command('activity')
@@ -146,15 +137,6 @@ def add_recommendation(controller, title, description):
 @handle_api_error
 def describe_activity(controller, days, view, tests, tags, endpoints, dos, statuses, convert):
     """ View my Detect results """
-
-    # setup conversion
-
-    build = BuildController(account=controller.account)
-    my_tests = {row['id']: row['name'] for row in build.list_tests()}
-    conversion = lambda i: my_tests.get(i, 'DELETED') if convert else i
-
-    # establish filters
-
     filters = dict(
         start=datetime.combine(datetime.utcnow() - timedelta(days=days), time.min),
         finish=datetime.combine(datetime.utcnow(), time.max)
@@ -168,80 +150,5 @@ def describe_activity(controller, days, view, tests, tags, endpoints, dos, statu
     if dos:
         filters['dos'] = dos
 
-    # build reports
-
-    raw = controller.describe_activity(view=view, filters=filters)
-    report = Table()
-
-    if view == 'logs':
-        report.add_column('timestamp')
-        report.add_column('test')
-        report.add_column('endpoint')
-        report.add_column('status')
-
-        raw.reverse()
-        for record in raw:
-            report.add_row(
-                record['date'], 
-                conversion(record['test']),
-                record['endpoint_id'], 
-                ExitCode(record['status']).name
-            )
-
-    elif view == 'insights':
-        report.add_column('dos')
-        report.add_column('test')
-        report.add_column('protected', style='green')
-        report.add_column('unprotected',  style='red')
-        report.add_column('error', style='yellow')
+    print_json(data=controller.describe_activity(view=view, filters=filters))
     
-        for ins in raw:
-            vol = ins['volume']
-            report.add_row(
-                ins['dos'], 
-                conversion(ins['test']), 
-                str(vol["protected"]),
-                str(vol["unprotected"]),
-                str(vol["error"])
-            )
-
-    elif view == 'probes':
-        report.add_column('endpoint_id')
-        report.add_column('dos')
-        report.add_column('state')
-        report.add_column('tags')
-
-        for ep in raw:
-            tags = ",".join(ep.get('tags'))
-            report.add_row(ep.get('endpoint_id'), ep.get('dos'), ep.get('state'), tags)
-
-    elif view == 'days':
-        report.add_column('date')
-        report.add_column('unprotected',  style='red')
-        report.add_column('volume', style='green')
-
-        for day in raw:
-            report.add_row(
-                day.get('date'), 
-                str(day.get('unprotected', 0)), 
-                str(day.get('count', 0))
-            )
-
-    elif view == 'rules':
-        report.add_column('VSR')
-        report.add_column('rule')
-        report.add_column('unprotected', style='red')
-        report.add_column('volume', style='green')
-
-        for entry in raw:
-            rule = entry.get('rule')
-            usage = entry.get('usage')
-            report.add_row(
-                rule.get('id'),
-                rule.get('label').replace('_', ' ').lower(),
-                str(usage.get('unprotected', 0)), 
-                str(usage.get('count', 0))
-            )
-
-    console = Console()
-    console.print(report)
