+++ python/sdk/prelude_sdk/controllers/detect_controller.py
@@ -56,6 +56,15 @@ class DetectController:
             raise Exception(res.text)
 
     @verify_credentials
+    def list_tests(self):
+        """ List all tests available to an account """
+        with Spinner():
+            res = requests.get(f'{self.account.hq}/build/tests', headers=self.account.headers)
+            if res.status_code == 200:
+                return res.json()
+            raise Exception(res.text)
+
+    @verify_credentials
     def enable_test(self, ident: str, run_code: int, tags: str):
         """ Enable a test so endpoints will start running it """
         with Spinner():
